fixing state of the chunk when multiple moves on edge done
As we simulate a go back to the beginning of the game each time the board is expanded, the state was also restore to the initial context / state. That does not interest us in the case we just want to get back to beginning with a new board size, to then apply ourselves the move made, mapped to the new board.