Mapping chunks from initial size to current size
Before, current chunks were kept the same and just mapped to the new board. Now chunks are reset to initial size of the board, as the moves re-applied will automatically update the chunks properly. So chunks need to be mapped from init size to new size, unlike before which was mapped between current size and new size. Yet, when re-applying the moves, they are mapped from current to new, and not init to new.

WIP : need to find a way to run the test - as I am not able to reset the chunks without using PlayerDesktop yet