update topology before reseting game when roll back
Reseting the game in a boardless game will check the center of the plate to reseting properly (see resetStateTo), this means the center of the plate should be the one of the new plate. But the remake trial should still be applied after the game has been reset. This was really problematic only when going back on the first move done on an edge